moved some compute implementations temporarily before they are refactored into proper layer implementations. introduced cached layer, which will forward the array type
