+++ source/layer/algorithm/normal.hpp
