+++ source/layer/constant.hpp
@@ -62,7 +62,9 @@ struct constant {
   T operator()(const size_t index){
     return typeselect(this->type(),
       [self=this, index]<typename S>() -> T {
-        if constexpr (std::convertible_to<S, T>){
+        if constexpr (std::same_as<T, S>){
+        return self->as<S>().operator()(index);
+        } else if constexpr (std::convertible_to<S, T>){
           return (T)self->as<S>().operator()(index);
         } else throw soil::error::cast_error<S, T>{}();
       }
@@ -72,7 +74,9 @@ struct constant {
   template<typename T>
   static typedbase* make(const soil::dtype type, const T value){
     return typeselect(type, [value]<typename S>() -> typedbase* {
-      if constexpr (std::convertible_to<T, S>){
+      if constexpr (std::same_as<T, S>){
+        return new soil::constant_t<S>(value);
+      } else if constexpr (std::convertible_to<T, S>){
         return new soil::constant_t<S>(S(value));
       } else throw soil::error::cast_error<T, S>{}();
     });
